// src/data/gymSpritesGenerated.ts
export interface GymTrainerSprite {
  graphicsId: string;
  spriteSheet: string;
  spriteWidth: number;
  spriteHeight: number;
  spriteFrame: number;
}